refactor(website): move asset filters to shared queries module
Relocate filterPlayableHeroes and filterShopableItems from the
deadlockdle-local queries module to app/queries/asset-queries.ts so
the flashcards routes and other consumers can share them without
reaching into a feature directory.
