+++ website/app/queries/asset-queries.ts
@@ -1,5 +1,5 @@
 import { queryOptions } from "@tanstack/react-query";
-import type { AbilityV2, UpgradeV2 } from "assets_deadlock_api_client/api";
+import type { AbilityV2, HeroV2, UpgradeV2 } from "assets_deadlock_api_client/api";
 
 import { CACHE_DURATIONS } from "~/constants/cache";
 import { assetsApi } from "~/lib/assets-api";
@@ -36,3 +36,11 @@ export const abilitiesQueryOptions = queryOptions({
   },
   staleTime: CACHE_DURATIONS.FOREVER,
 });
+
+export function filterPlayableHeroes(heroes: HeroV2[]): HeroV2[] {
+  return heroes.filter((h) => h.player_selectable && !h.disabled && !h.in_development);
+}
+
+export function filterShopableItems(items: UpgradeV2[]): UpgradeV2[] {
+  return items.filter((item) => item.shopable && !item.disabled && item.shop_image_webp);
+}
