+++ website/app/routes/flashcards.heroes.tsx
@@ -4,9 +4,8 @@ import { useMemo } from "react";
 import type { MetaFunction } from "react-router";
 
 import { createPageMeta } from "~/lib/meta";
-import { heroesQueryOptions } from "~/queries/asset-queries";
+import { filterPlayableHeroes, heroesQueryOptions } from "~/queries/asset-queries";
 
-import { filterPlayableHeroes } from "./deadlockdle/lib/queries";
 import { FlashcardGame } from "./flashcards/components/FlashcardGame";
 
 export const meta: MetaFunction = () => {
